rename packages from "clocker" to "onthefly"
