+++ software/base/src/test/java/org/apache/brooklyn/core/location/dynamic/onthefly/StubInfrastructure.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.brooklyn.core.location.dynamic.clocker;
+package org.apache.brooklyn.core.location.dynamic.onthefly;
 
 import java.util.List;
