+++ software/base/src/test/java/org/apache/brooklyn/core/location/dynamic/onthefly/StubResolver.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.brooklyn.core.location.dynamic.clocker;
+package org.apache.brooklyn.core.location.dynamic.onthefly;
 
 import static com.google.common.base.Preconditions.checkNotNull;
